dl config api update
