+++ catalyst/dl/__init__.py
@@ -3,7 +3,6 @@
 
 from catalyst.core import *
 
-from .core import *
 from .callbacks import *
 from .experiment import *
 from .state import *
