+++ catalyst/dl/callbacks/metrics/f1_score.py
@@ -1,4 +1,4 @@
-from catalyst.dl.core import MetricCallback
+from catalyst.core import MetricCallback
 from catalyst.utils import metrics
