--- catalyst/dl/core/__init__.py
@@ -1,7 +1,0 @@
-# flake8: noqa
-# isort:skip_file
-
-from catalyst.core import *
-from .callback import MeterMetricsCallback
-from .experiment import Experiment
-from .runner import Runner
