+++ catalyst/dl/experiment/__init__.py
@@ -1,8 +1,6 @@
 # flake8: noqa
 
-from .base import BaseExperiment
 from .config import ConfigExperiment
+from .core import Experiment
 from .gan import GanExperiment
 from .supervised import SupervisedExperiment
-
-from catalyst.contrib.dl.experiment import *  # isort:skip
