+++ catalyst/dl/experiment/supervised.py
@@ -4,28 +4,20 @@ from torch.optim.lr_scheduler import ReduceLROnPlateau
 
 from catalyst.dl import (
     Callback,
-    CheckpointCallback,
-    CheckRunCallback,
-    ConsoleLogger,
     CriterionCallback,
-    ExceptionCallback,
-    MetricManagerCallback,
     OptimizerCallback,
     SchedulerCallback,
-    TensorboardLogger,
-    TimerCallback,
-    ValidationManagerCallback,
-    VerboseLogger,
 )
 from catalyst.utils.tools.typing import Criterion, Optimizer, Scheduler
 
-from .base import BaseExperiment
+from .core import Experiment
 
 
-class SupervisedExperiment(BaseExperiment):
-    """Supervised experiment.
+class SupervisedExperiment(Experiment):
+    """
+    Supervised experiment.
 
-    The main difference with BaseExperiment that it will
+    The main difference with Experiment that it will
     add several callbacks by default if you haven't.
 
     Here are list of callbacks by default:
@@ -59,15 +51,12 @@ class SupervisedExperiment(BaseExperiment):
                 training stages.
 
         Returns:
-            List[Callback]: list of callbacks for experiment
+            (OrderedDict[str, Callback]): Ordered dictionary of callbacks
+                for experiment
         """
-        callbacks = self._callbacks or OrderedDict()
+        callbacks = super().get_callbacks(stage=stage) or OrderedDict()
 
         default_callbacks = []
-        if self._verbose:
-            default_callbacks.append(("_verbose", VerboseLogger))
-        if self._check_run:
-            default_callbacks.append(("_check", CheckRunCallback))
 
         if not stage.startswith("infer"):
             if self._criterion is not None and isinstance(
@@ -82,16 +71,6 @@ class SupervisedExperiment(BaseExperiment):
                 self._scheduler, (Scheduler, ReduceLROnPlateau)
             ):
                 default_callbacks.append(("_scheduler", SchedulerCallback))
-
-            default_callbacks.append(("_timer", TimerCallback))
-            default_callbacks.append(("_metrics", MetricManagerCallback))
-            default_callbacks.append(
-                ("_validation", ValidationManagerCallback)
-            )
-            default_callbacks.append(("_saver", CheckpointCallback))
-            default_callbacks.append(("_console", ConsoleLogger))
-            default_callbacks.append(("_tensorboard", TensorboardLogger))
-        default_callbacks.append(("_exception", ExceptionCallback))
 
         for callback_name, callback_fn in default_callbacks:
             is_already_present = any(
