+++ catalyst/dl/runner/__init__.py
@@ -1,6 +1,5 @@
 # flake8: noqa
 
+from .core import Runner
 from .gan import GanRunner, MultiPhaseRunner
 from .supervised import SupervisedRunner
-
-from catalyst.contrib.dl.runner import *  # isort:skip
