--- tests/_tests_scripts/_alchemy.py
@@ -1,2 +1,0 @@
-# flake8: noqa
-from catalyst.dl import AlchemyRunner
