--- tests/_tests_scripts/dl_core.py
@@ -1,2 +1,0 @@
-# flake8: noqa
-from catalyst.dl.core import *
